Create folder noti in folder components and add swal and noti in BoxMenuUser.tsx
