+++ src/pages/Admin-page/Admin.css
@@ -236,6 +236,7 @@
 .user-wrapper img {
   border-radius: 50%;
   margin-right: 1rem;
+  margin-left: 2rem;
 }
 
 .user-wrapper:hover .menu-box {
