+++ src/pages/User-page/components/BoxMenuUser.tsx
@@ -1,19 +1,12 @@
-import {
-  avatar,
-  box,
-  adress,
-  voucher,
-  comment,
-  iblog,
-  watched,
-} from "../../../import/import-assets";
+import { avatar, box, adress, voucher, comment, iblog, watched } from "../../../import/import-assets";
 import { getNameFromToken } from "../../../utils/jwtHelper";
-import { useMemo } from "../../../import/import-another";
+import { useMemo, swal2 } from "../../../import/import-another";
 import { Link } from "../../../import/import-libary";
-
 
 const BoxMenuUser = () => {
   const token = localStorage.getItem("token");
+
+//---------------- Get User Name ------------------------------------------------
 
   const userName = useMemo(() => {
     if (!token) {
@@ -25,6 +18,19 @@ const BoxMenuUser = () => {
     return usernameIdentifier;
   }, [token]);
 
+//---------------- Handle Click Unavailable Feature ------------------------------
+
+  const handleClickUnavailableFeature = () => {
+    swal2.fire({
+      icon: "info",
+      title: "Feature Unavailable",
+      text: "This feature is currently being updated. Please check back later!",
+      timer: 3000,
+      timerProgressBar: true,
+      showConfirmButton: false,
+    });
+  };
+
   return (
     <div className="box-menu-user">
       <div className="box-profile-avatar">
@@ -33,9 +39,11 @@ const BoxMenuUser = () => {
         </div>
         <div className="onlick-profile">
           <h5>{userName}</h5>
-          <p>
-            <Link to="/profile">Profile user</Link>
-          </p>
+          <Link to="/profile" className="profile-user">
+            <p>
+              Profile user {`>>`}
+            </p>
+          </Link>
         </div>
       </div>
       <div className="box-oder-adress-voucher">
@@ -43,26 +51,29 @@ const BoxMenuUser = () => {
           <img src={box} alt="" />
           <p>Purchase order</p>
         </Link>
-        <div className="my-adress">
+        <div className="my-adress" onClick={handleClickUnavailableFeature}>
           <img src={adress} alt="" />
           <p>Address</p>
         </div>
-        <div className="my-voucher">
+        <div className="my-voucher" onClick={handleClickUnavailableFeature}>
           <img src={voucher} alt="" />
           <p>My Voucher</p>
         </div>
       </div>
 
       <div className="box-oder-adress-voucher">
-        <div className="purchased-order">
+        <div
+          className="purchased-order"
+          onClick={handleClickUnavailableFeature}
+        >
           <img src={comment} alt="" />
           <p>My review</p>
         </div>
-        <div className="my-adress">
+        <div className="my-adress" onClick={handleClickUnavailableFeature}>
           <img src={iblog} alt="" />
           <p>blog viewed</p>
         </div>
-        <div className="my-voucher">
+        <div className="my-voucher" onClick={handleClickUnavailableFeature}>
           <img src={watched} alt="" />
           <p>Viewed products</p>
         </div>
